+++ crates/web-client/js/types/index.d.ts
@@ -1,5 +1,8 @@
 export {
+  Account, 
+  AccountHeader,
   AccountId,
+  AccountStorageMode,
   AdviceMap,
   AuthSecretKey,
   Felt,